git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@399 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/DPD/atom_vec_dpd.h
@@ -22,11 +22,11 @@ class AtomVecDPD : public AtomVecAtomic {
  public:
   AtomVecDPD(class LAMMPS *, int, char **);
   void zero_ghost(int, int);
-  int pack_comm(int, int *, double *, int, double *);
+  int pack_comm(int, int *, double *, int, int *);
   int pack_comm_one(int, double *);
   void unpack_comm(int, int, double *);
   int unpack_comm_one(int, double *);
-  int pack_border(int, int *, double *, int, double *);
+  int pack_border(int, int *, double *, int, int *);
   int pack_border_one(int, double *);
   void unpack_border(int, int, double *);
   int unpack_border_one(int, double *);
