+++ src/MANYBODY/pair_eam.h
@@ -31,7 +31,7 @@ class PairEAM : public Pair {
 
   void single_embed(int, int, double &);
 
-  int pack_comm(int, int *, double *, int, double *);
+  int pack_comm(int, int *, double *, int, int *);
   void unpack_comm(int, int, double *);
   int pack_reverse_comm(int, int, double *);
   void unpack_reverse_comm(int, int *, double *);
