+++ src/MOLECULE/atom_vec_angle.h
@@ -26,11 +26,11 @@ class AtomVecAngle : public AtomVec {
   void zero_owned(int);
   void zero_ghost(int, int);
   void copy(int, int);
-  int pack_comm(int, int *, double *, int, double *);
+  int pack_comm(int, int *, double *, int, int *);
   void unpack_comm(int, int, double *);
   int pack_reverse(int, int, double *);
   void unpack_reverse(int, int *, double *);
-  int pack_border(int, int *, double *, int, double *);
+  int pack_border(int, int *, double *, int, int *);
   int pack_border_one(int, double *);
   void unpack_border(int, int, double *);
   int unpack_border_one(int, double *);
