+++ src/fix_shake.h
@@ -34,7 +34,7 @@ class FixShake : public Fix {
   void copy_arrays(int, int);
   int pack_exchange(int, double *);
   int unpack_exchange(int, double *);
-  int pack_comm(int, int *, double *, int, double *);
+  int pack_comm(int, int *, double *, int, int *);
   void unpack_comm(int, int, double *);
 
   int dof(int);
