+++ src/pair.h
@@ -79,7 +79,7 @@ class Pair : protected Pointers {
   virtual void write_restart_settings(FILE *) {}
   virtual void read_restart_settings(FILE *) {}
 
-  virtual int pack_comm(int, int *, double *, int, double *) {return 0;}
+  virtual int pack_comm(int, int *, double *, int, int *) {return 0;}
   virtual void unpack_comm(int, int, double *) {}
   virtual int pack_reverse_comm(int, int, double *) {return 0;}
   virtual void unpack_reverse_comm(int, int *, double *) {}
